[dagit] Remove flat left nav
