+++ js_modules/dagit/packages/core/src/app/Flags.tsx
@@ -8,7 +8,6 @@ export const DAGIT_FLAGS_KEY = 'DAGIT_FLAGS';
 export enum FeatureFlag {
   flagDebugConsoleLogging = 'flagDebugConsoleLogging',
   flagDisableWebsockets = 'flagDisableWebsockets',
-  flagFlatLeftNav = 'flagFlatLeftNav',
 }
 
 export const getFeatureFlags: () => FeatureFlag[] = memoize(
