+++ js_modules/dagit/packages/core/src/app/UserSettingsRoot.tsx
@@ -114,16 +114,6 @@ const UserSettingsRoot: React.FC<SettingsRootProps> = ({tabs}) => {
                 />
               ),
             },
-            {
-              key: 'Sectioned left nav (opt-out)',
-              value: (
-                <Checkbox
-                  format="switch"
-                  checked={!flags.includes(FeatureFlag.flagFlatLeftNav)}
-                  onChange={() => toggleFlag(FeatureFlag.flagFlatLeftNav)}
-                />
-              ),
-            },
           ]}
         />
       </Box>
